+++ src/aux-common/runtime/Utils.ts
@@ -7,6 +7,7 @@ import {
     hasValue,
     isBot,
     isRuntimeBot,
+    ORIGINAL_OBJECT,
 } from '../bots/BotCalculations';
 import { AuxPartitionRealtimeStrategy } from '../partitions/AuxPartition';
 import { forOwn } from 'lodash';
@@ -472,6 +473,12 @@ export function ensureBotIsSerializable(bot: Bot): Bot {
             value instanceof Rotation
         ) {
             updateTag(tag, convertToCopiableValue(value));
+        } else if (
+            hasValue(value) &&
+            typeof value === 'object' &&
+            ORIGINAL_OBJECT in value
+        ) {
+            updateTag(tag, value[ORIGINAL_OBJECT]);
         }
     }
 
@@ -483,4 +490,43 @@ export function ensureBotIsSerializable(bot: Bot): Bot {
         }
         newBot.tags[tag] = value;
     }
+}
+
+/**
+ * Determines if the given value represents a promise.
+ * @param value The value to check.
+ */
+export function isPromise<T>(value: unknown): value is Promise<T> {
+    return (
+        typeof value === 'object' &&
+        value !== null &&
+        value instanceof Promise &&
+        typeof (value as any).then === 'function' &&
+        typeof (value as any).catch === 'function'
+    );
+}
+
+export const RUNTIME_PROMISE = Symbol('RUNTIME_PROMISE');
+
+export interface RuntimePromise<T> extends Promise<T> {
+    [RUNTIME_PROMISE]: true;
+}
+
+export function isRuntimePromise<T>(
+    value: unknown
+): value is RuntimePromise<T> {
+    return isPromise(value) && RUNTIME_PROMISE in value;
+}
+
+export function markAsRuntimePromise<T>(
+    promise: Promise<T>
+): RuntimePromise<T> {
+    Object.defineProperty(promise, RUNTIME_PROMISE, {
+        value: true,
+        configurable: false,
+        enumerable: false,
+        writable: false,
+    });
+
+    return promise as RuntimePromise<T>;
 }