+++ src/aux-server/aux-web/shared/interaction/CameraControls.ts
@@ -818,6 +818,7 @@ export class CameraControls {
                     Math.sqrt(startDir.lengthSq() * endDir.lengthSq());
                 xAngle *= this.rotateSpeed;
                 xAngle = Math.acos(ThreeMath.clamp(xAngle, -1, 1));
+
                 let cross = startDir.x * endDir.y - startDir.y * endDir.x;
                 if (cross >= 0) xAngle = -xAngle;
                 this.rotateLeft(xAngle);
@@ -918,31 +919,44 @@ export class CameraControls {
                     this.touchRotateStart.finger0 = input.getTouchPagePos(0);
                     this.touchRotateStart.finger1 = input.getTouchPagePos(1);
 
-                    const finger0 = input.getTouchScreenPos(0);
-                    const finger1 = input.getTouchScreenPos(1);
+                    if (this._camera instanceof PerspectiveCamera) {
+                        this.originMidpoint = null;
+                    } else {
+                        const finger0 = input.getTouchScreenPos(0);
+                        const finger1 = input.getTouchScreenPos(1);
 
-                    const originFinger0Ray = Physics.rayAtScreenPos(
-                        finger0,
-                        this._camera
-                    );
-                    const originFinger1Ray = Physics.rayAtScreenPos(
-                        finger1,
-                        this._camera
-                    );
-                    const originFinger0Point = Physics.pointOnPlane(
-                        originFinger0Ray,
-                        new Plane(new Vector3(0, 1, 0))
-                    );
-                    const originFinger1Point = Physics.pointOnPlane(
-                        originFinger1Ray,
-                        new Plane(new Vector3(0, 1, 0))
-                    );
+                        const originFinger0Ray = Physics.rayAtScreenPos(
+                            finger0,
+                            this._camera
+                        );
+                        const originFinger1Ray = Physics.rayAtScreenPos(
+                            finger1,
+                            this._camera
+                        );
+                        const originFinger0Point = Physics.pointOnPlane(
+                            originFinger0Ray,
+                            new Plane(new Vector3(0, 0, 1))
+                        );
+                        const originFinger1Point = Physics.pointOnPlane(
+                            originFinger1Ray,
+                            new Plane(new Vector3(0, 0, 1))
+                        );
 
-                    this.originMidpoint = new Vector3(
-                        (originFinger0Point.x + originFinger1Point.x) / 2,
-                        (originFinger0Point.y + originFinger1Point.y) / 2,
-                        (originFinger0Point.z + originFinger1Point.z) / 2
-                    );
+                        if (originFinger0Point && originFinger1Point) {
+                            this.originMidpoint = new Vector3(
+                                (originFinger0Point.x + originFinger1Point.x) /
+                                    2,
+                                (originFinger0Point.y + originFinger1Point.y) /
+                                    2,
+                                (originFinger0Point.z + originFinger1Point.z) /
+                                    2
+                            );
+                        } else if (originFinger0Point) {
+                            this.originMidpoint = originFinger0Point.clone();
+                        } else {
+                            this.originMidpoint = originFinger1Point.clone();
+                        }
+                    }
 
                     this.state = STATE.TOUCH_ROTATE_ZOOM;
                 }
@@ -1246,7 +1260,10 @@ export class CameraControls {
         ) {
             rotationTarget = this.originMidpoint;
 
-            lookTarget.sub(this.originMidpoint);
+            rotationTarget.applyQuaternion(quat);
+            lookTarget.applyQuaternion(quat);
+
+            lookTarget.sub(rotationTarget);
             const s = new Spherical().setFromVector3(lookTarget);
 
             s.theta += this.sphericalDelta.theta;
@@ -1258,7 +1275,10 @@ export class CameraControls {
 
             lookTarget.setFromSpherical(s);
 
-            lookTarget.add(this.originMidpoint);
+            lookTarget.add(rotationTarget);
+
+            rotationTarget.applyQuaternion(quatInverse);
+            lookTarget.applyQuaternion(quatInverse);
         }
 
         offset.copy(position).sub(rotationTarget);
@@ -1323,6 +1343,7 @@ export class CameraControls {
             // move target to panned location
             this.target.add(this.panOffset);
             this.target.add(this.cameraFrameOffset);
+
             if (this.cameraFrameOffset.length() > 0) {
                 this.currentDistX = this.target.x;
                 this.currentDistY = this.target.y;