+++ src/aux-server/typings/global.d.ts
@@ -8,7 +8,22 @@ declare module 'filepond-plugin-file-validate-type';
 declare module 'jsbarcode';
 declare module 'clipboard-polyfill';
 declare module 'rollup-plugin-copy';
-declare module 'virtual:pwa-register';
+
+declare module 'virtual:pwa-register' {
+    export interface RegisterSWOptions {
+        immediate?: boolean;
+        onNeedRefresh?: () => void;
+        onOfflineReady?: () => void;
+        onRegistered?: (
+            registration: ServiceWorkerRegistration | undefined
+        ) => void;
+        onRegisterError?: (error: any) => void;
+    }
+
+    export function registerSW(
+        options?: RegisterSWOptions
+    ): (reloadPage?: boolean) => Promise<void>;
+}
 // declare module 'monaco-editor-core';
 
 declare module '*.jpg' {